refactor: separate StoryOperations::UpdateAll into its own file
